refactor: split a test file
